move all the attributes to the root Ministry object
