+++ app/src/main/java/com/expidev/gcmapp/model/AssociatedMinistry.java
@@ -1,15 +1,12 @@
 package com.expidev.gcmapp.model;
 
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.EnumSet;
 import java.util.List;
 
 /**
@@ -24,15 +21,6 @@ public class AssociatedMinistry extends Ministry
     public static final String JSON_LATITUDE = "latitude";
     public static final String JSON_LONGITUDE = "longitude";
     public static final String JSON_LOCATION_ZOOM = "location_zoom";
-
-    @Nullable
-    private String parentMinistryId;
-    private String ministryCode;
-    @NonNull
-    private final EnumSet<Mcc> mccs = EnumSet.noneOf(Mcc.class);
-    private double latitude;
-    private double longitude;
-    private int locationZoom;
 
     @NonNull
     public static List<AssociatedMinistry> listFromJson(@NonNull final JSONArray json) throws JSONException {
@@ -68,120 +56,5 @@ public class AssociatedMinistry extends Ministry
         ministry.setLocationZoom(json.optInt(JSON_LOCATION_ZOOM));
 
         return ministry;
-    }
-
-    @Nullable
-    public String getParentMinistryId() {
-        return parentMinistryId;
-    }
-
-    public void setParentMinistryId(@Nullable final String parentMinistryId) {
-        this.parentMinistryId = parentMinistryId;
-    }
-
-    public String getMinistryCode()
-    {
-        return ministryCode;
-    }
-
-    public void setMinistryCode(String ministryCode)
-    {
-        this.ministryCode = ministryCode;
-    }
-
-    @NonNull
-    public EnumSet<Mcc> getMccs() {
-        return EnumSet.copyOf(this.mccs);
-    }
-
-    public boolean hasMcc(@NonNull final Mcc mcc) {
-        return this.mccs.contains(mcc);
-    }
-
-    public void setMccs(final Collection<Mcc> mccs) {
-        this.mccs.clear();
-        if (mccs != null) {
-            this.mccs.addAll(mccs);
-        }
-    }
-
-    public boolean hasSlm()
-    {
-        return hasMcc(Mcc.SLM);
-    }
-
-    public void setHasSlm(boolean hasSlm)
-    {
-        if (hasSlm) {
-            this.mccs.add(Mcc.SLM);
-        } else {
-            this.mccs.remove(Mcc.SLM);
-        }
-    }
-
-    public boolean hasLlm()
-    {
-        return hasMcc(Mcc.LLM);
-    }
-
-    public void setHasLlm(boolean hasLlm)
-    {
-        if (hasLlm) {
-            this.mccs.add(Mcc.LLM);
-        } else {
-            this.mccs.remove(Mcc.LLM);
-        }
-    }
-
-    public boolean hasDs()
-    {
-        return hasMcc(Mcc.DS);
-    }
-
-    public void setHasDs(boolean hasDs)
-    {
-        if (hasDs) {
-            this.mccs.add(Mcc.DS);
-        } else {
-            this.mccs.remove(Mcc.DS);
-        }
-    }
-
-    public boolean hasGcm()
-    {
-        return hasMcc(Mcc.GCM);
-    }
-
-    public void setHasGcm(boolean hasGcm)
-    {
-        if (hasGcm) {
-            this.mccs.add(Mcc.GCM);
-        } else {
-            this.mccs.remove(Mcc.GCM);
-        }
-    }
-
-    public double getLatitude() {
-        return latitude;
-    }
-
-    public void setLatitude(final double latitude) {
-        this.latitude = latitude;
-    }
-
-    public double getLongitude() {
-        return longitude;
-    }
-
-    public void setLongitude(final double longitude) {
-        this.longitude = longitude;
-    }
-
-    public int getLocationZoom() {
-        return locationZoom;
-    }
-
-    public void setLocationZoom(final int locationZoom) {
-        this.locationZoom = locationZoom;
     }
 }
