+++ app/src/main/java/com/expidev/gcmapp/model/Ministry.java
@@ -4,11 +4,10 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
 import java.io.Serializable;
+import java.util.Collection;
+import java.util.EnumSet;
 import java.util.Locale;
 
-/**
- * Created by William.Randall on 1/9/2015.
- */
 public class Ministry extends Base implements Serializable
 {
     private static final long serialVersionUID = 0L;
@@ -44,7 +43,15 @@ public class Ministry extends Base implements Serializable
 
     @NonNull
     private String ministryId = INVALID_ID;
+    @Nullable
+    private String parentMinistryId;
+    private String ministryCode;
     private String name;
+    @NonNull
+    private final EnumSet<Mcc> mccs = EnumSet.noneOf(Mcc.class);
+    private double latitude;
+    private double longitude;
+    private int locationZoom;
 
     @NonNull
     public String getMinistryId()
@@ -57,6 +64,23 @@ public class Ministry extends Base implements Serializable
         this.ministryId = ministryId;
     }
 
+    @Nullable
+    public String getParentMinistryId() {
+        return parentMinistryId;
+    }
+
+    public void setParentMinistryId(@Nullable final String parentMinistryId) {
+        this.parentMinistryId = parentMinistryId;
+    }
+
+    public String getMinistryCode() {
+        return ministryCode;
+    }
+
+    public void setMinistryCode(final String ministryCode) {
+        this.ministryCode = ministryCode;
+    }
+
     public String getName()
     {
         return name;
@@ -65,6 +89,94 @@ public class Ministry extends Base implements Serializable
     public void setName(String name)
     {
         this.name = name;
+    }
+
+    @NonNull
+    public EnumSet<Mcc> getMccs() {
+        return EnumSet.copyOf(this.mccs);
+    }
+
+    public boolean hasMcc(@NonNull final Mcc mcc) {
+        return this.mccs.contains(mcc);
+    }
+
+    public void setMccs(final Collection<Mcc> mccs) {
+        this.mccs.clear();
+        if (mccs != null) {
+            this.mccs.addAll(mccs);
+        }
+    }
+
+    public boolean hasSlm() {
+        return hasMcc(Mcc.SLM);
+    }
+
+    public void setHasSlm(boolean hasSlm) {
+        if (hasSlm) {
+            this.mccs.add(Mcc.SLM);
+        } else {
+            this.mccs.remove(Mcc.SLM);
+        }
+    }
+
+    public boolean hasLlm() {
+        return hasMcc(Mcc.LLM);
+    }
+
+    public void setHasLlm(boolean hasLlm) {
+        if (hasLlm) {
+            this.mccs.add(Mcc.LLM);
+        } else {
+            this.mccs.remove(Mcc.LLM);
+        }
+    }
+
+    public boolean hasDs() {
+        return hasMcc(Mcc.DS);
+    }
+
+    public void setHasDs(boolean hasDs) {
+        if (hasDs) {
+            this.mccs.add(Mcc.DS);
+        } else {
+            this.mccs.remove(Mcc.DS);
+        }
+    }
+
+    public boolean hasGcm() {
+        return hasMcc(Mcc.GCM);
+    }
+
+    public void setHasGcm(boolean hasGcm) {
+        if (hasGcm) {
+            this.mccs.add(Mcc.GCM);
+        } else {
+            this.mccs.remove(Mcc.GCM);
+        }
+    }
+
+    public double getLatitude() {
+        return latitude;
+    }
+
+    public void setLatitude(final double latitude) {
+        this.latitude = latitude;
+    }
+
+    public double getLongitude() {
+        return longitude;
+    }
+
+    public void setLongitude(final double longitude) {
+        this.longitude = longitude;
+    }
+
+    public int getLocationZoom() {
+        return locationZoom;
+    }
+
+    public void setLocationZoom(final int locationZoom) {
+        this.locationZoom = locationZoom;
     }
 
     @Override
